Remove dead code. Don't call into printf().
There already seems to be an assertion in place that s->error_proc
cannot be NULL. There are also various other parts of code that call
into this function without checking for NULL first.

The intent of this change is that it makes this code build with
Nuxi CloudABI[1]. CloudABI is a sandboxed runtime computing environment.
As it is mainly oriented towards running (networked) services, it does
not understand the concept of stdin and stdout.

[1] Nuxi CloudABI: https://github.com/NuxiNL/cloudlibc