Qualify "newline at EOF not required" with actually being there
in C++11 and later modes Wave by default does not require newline
prior to EOF. This works fine but there is one place where the test
for this feature was used where Wave is not necessarily looking at the
end of the file: pp_is_last_on_line. One observable (and surprising)
result is that unknown directives are not getting flagged, but there
may be others. The unit tests did not discover this problem because
they generally do not set c++11 mode.