Extend __has_include grammar for complex conditional expressions
The logic handling __has_include erroneously assumed it would always
be the last expression on a line. This code fixes that by extending
the grammar to only consume the __has_include itself(), and count
parentheses as necessary to find the correct input range.

Also add some unit tests to cover the error in the future.