Suppress gcc 11.2 -Wfree-nonheap-object warnings.
The warnings are bogus because the pData_ pointer never points to
emptyString_ if capacity() == 0, and always points otherwise. The
compiler simply fails to deduce this invariant.

Unfortunately, suppressing the warning with a #pragma doesn't work,
so we have to invent a compile-time assert that will make the invariant
detectable by the compiler.

Closes https://github.com/boostorg/wave/issues/159.