Fix warnings about freeing non-heap-allocated object (#163)
* Use capacity() == 0 as the definitive mark of an empty string storage.

Comparing the pointer with the static emptyString_ member of the
SimpleStringStorage class will likely break if the comparison happens
across shared library boundary. So use capacity() == 0 consistently
to detect whether the storage refers to the emptyString_ member.
When capacity is non-zero the data is always dynamically allocated.

* Suppress gcc 11.2 -Wfree-nonheap-object warnings.

The warnings are bogus because the pData_ pointer never points to
emptyString_ if capacity() == 0, and always points otherwise. The
compiler simply fails to deduce this invariant.

Unfortunately, suppressing the warning with a #pragma doesn't work,
so we have to invent a compile-time assert that will make the invariant
detectable by the compiler.

Closes https://github.com/boostorg/wave/issues/159.

* Removed unused AllocatorStringStorage::Realloc function.

This function was not used and potentially incorrect, as it unconditionally
calls Free, which would be wrong if the original buffer was emptyString_.