Check for backslash characters before the start of the new data
With BOOST_WAVE_BSIZE set to 40, test t_5_002 fails because a newline
appears exactly at the beginning of a newly fetched buffer, and a
backslash is at the end of the previous one. In that case the escaped
newline was not detected. This change will also consider three
unprocessed bytes of input data, if available, prior to the new data,
which is enough to detect a trigraph backslash.